Creamos el modulo de transacciones, permitiendo realizar: - retiros - depositos - transferencias - consulta de las transacciones dependiendo el flitro aplicado

+++ src/main/java/com/xarhabia/BankApp/account/dto/response/AccountResponse.java
@@ -1,10 +1,11 @@
 package com.xarhabia.BankApp.account.dto.response;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record AccountResponse(
         String accountNumber,
-        double balance,
+        BigDecimal balance,
         String accountType,
         String description,
         OffsetDateTime createAt
